feat(ckdoge): derive Dogecoin address for deposits (#7066)
Use public key derivation to derive a Dogecoin address from a ledger
account that contains a user principal and an optional subaccount,
similarly to what is done in the ckBTC minter.

One difficulty is that the `Bitcoin` address type is used as a field in
`RetrieveBtcRequest`, which is both used in the minter event log (event
`AcceptedRetrieveBtcRequest`) and in the minter state
(`pending_retrieve_btc_requests`). Generalizing this to be able to deal
with different address types without impacting ckBTC is not trivial:

- a separate event log would be needed;
- and, the state would need to be made generic, requiring either dynamic
dispatch or feature flags (the state is held in a `thread_local!`,
implying that the type must be known at compile time)

To be able to reuse the same infrastructure, the 20 bytes of the
Dogecoin address is mapped to the corresponding 20-bytes of a Bitcoin
address before calling certain methods on the ckBTC minter.