feat(nns-recovery): [CON-1597] allow to upgrade without first electing the version (#7034)
This PR allows to upgrade to an already elected version in NNS
recoveries, which could be useful if we simply need to revert back to a
previous version (that is already elected).

We also add a new system test (`nr_no_bless`) that tests this feature.
Ticket CON-1599 will add a similar system test to application subnet
recoveries.