refactor(cmc/root): Fix a few places where ic_cdk_macros are not used (#7099)
# Why

There are a few places (root/cmc) not using the ic_cdk_macros without a
good reason.

# What

* Use `ic_cdk_macros::query(hidden = true)` for `total_cycles_minted`
* Switch `total_cycles_minted` to return candid rather than protobuf.
Note that the current value on the mainnet is at 70% of u64::MAX, so the
u64 doesn't make sense anyway. This is a breaking change, but the method
has never been exposed (not in .did)
* Use `#[init]` for root