chore: download directly from CDN in resuming_download_test (#7137)
This changes the `//rs/http_utils:resuming_download_test` to download
the update image directly from the CDN instead of from the
dc_http_proxy. That way this test can be run outside our infrastructure
like on Namespace which we're experimenting with in
https://github.com/dfinity/ic/pull/7121.

Since this test is a `long_test` it mostly runs only on master so it
doesn't generate too much traffic.