test(pocket-ic): test index and ledger with the same balance (#7142)
The broken test was compensating for the discrepancy in balance
calculation between the ledger and index.
https://github.com/dfinity/ic/pull/6508 fixed this discrepancy so both
index and ledger have the same balances. Since this PR is not on mainnet
yet, we temporarily have to compare against both correct balance (for
testing with master) and incorrect (for testing with mainnet).