feat(ICP_Rosetta): FI-1877: Add ic-agent timeout for initial sync (#7131)
Add a 1 second timeout to ic-agent to improve the performance of ICP
Rosetta's initial sync. With no timeout set, logs show that it is
possible to have frequent situations where there are 10-15s between
consecutive successful block retrievals by Rosetta from the ICP ledger
and/or archives, which can significantly increase the time it takes
Rosetta to sync, especially for an initial sync starting from scratch.
Setting a 1 second timeout resolves this.