test(http-oucalls): Add a test that checks that we can safely truncate error messages when the boundary is on a multi-byte character (#7157)
As of [enabling gossiping of full
response](https://github.com/dfinity/ic/pull/7092), we truncate error
messages to 1KB in order to protect our pool against OOM from malicious
replicas.

A naive
[truncate](https://doc.rust-lang.org/std/string/struct.String.html#method.truncate)
would panic the replica if the boundary was not perfectly aligned with a
character (multi-byte characters can cause this, such as emojis).

This test only asserts that we gracefully handle such a case.