feat: swapping functionality implementation (#7172)
This PR adds the functionality for swapping active nodes in a subnet
owned by single node operator. A node operator can swap one of its
assigned nodes in the same data center with an unassigned one.

At the moment this feature is completely turned off and won't be enabled
until system tests are added.