feat(cmc): support env vars in canister creation settings (#7082)
Add support for the canister setting `environment_variables` to
`create_canister` and `notify_create_canister` of the CMC.
This will enable the cycles ledger to also support environment variables
in canister settings when creating new canisters. This in turn will
enable `icp-cli` and maybe `dfx` to support these settings on canister
creation.