perf(idkg): CON-1593 Use multi-threading in `validate_dealing_support` (#7069)
This PR makes use of the IDKG thread pool to validate dealing support
shares in parallel. We iterate over all unvalidated support shares in
parallel, and attempt to validate each of them.

Note that we keep a cache of signers for which we have already validated
a support share for a given dealing. In order to be usable in the
multi-threaded setting, the cache is behind a `RwLock`.

Before processing an unvalidated share, we acquire the read lock on this
cache, and check if:
1. the same signer has already issued a support share for this dealing,
in which case the new share is invalidated
2. we already validated enough (2f+1) support shares for the
corresponding dealing, in which case the new share is removed
immediately.

If neither condition is true, we will release the read lock and continue
with cryptographic validation of the share.

Note that this introduces the following race condition: Two different
threads could each be assigned a different share by the same signer for
the same dealing. After reading the cache, both threads determine that
the share should be validated (despite the shares being duplicates of
each other).

Therefore, if the share turns out to be valid after cryptographic
verification, we acquire the _write_ lock on the cache, to insert the
new signer and check again if a share by the same signer was validated
in the meantime. If so, the newly verified share is invalidated instead.

### Alternatives Considered
An alternative approach (https://github.com/dfinity/ic/pull/7029) was to
iterate over the unvalidated shares twice (once sequentially, and once
in parallel): During the first iteration, we remove all duplicate shares
and select at most 2f+1 shares for each dealing to be cryptographically
verified as part of the second iteration. This would avoid the
aforementioned race condition and guarantee that no more than 2f+1
shares are validated (which is a performance optimization).

However, this approach turns out to be more complex, and potentially
insecure:

Assume we already verified 2f shares for a certain dealing and are now
only missing the last one. Then, during the sequential iteration, we
will only select a single share to be validated as part of the parallel
iteration. If this single share turns out to be invalid, we will select
a new share during the next call to `validate_dealing_support`. However,
since shares for the same transcript and dealing are sorted by the
signer's node ID, we will again pick a share by the same signer
(assuming they sent another one/multiple in the meantime). If this
signer turns out to be malicious, they could prevent us from ever
accumulating enough support shares.

Instead, using locks to guarantee exclusive write access to the cache is
simpler, and closer to the previous sequential implementation. The locks
are only acquired for a very short amount of time, which does not affect
performance according to performance tests.

In prinicpal, this approach could lead to frequently validating more
than 2f+1 shares. However, unvalidated shares are ordered by the
transcript ID. This means that under high load (where this optimization
matters), support shares for the same transcript ID should be assigned
to only one or two threads (modulo some work stealing), which should
reduce the number of race conditions and therefore the number of
unneeded validations.

---------

Co-authored-by: Pierugo Pace <pierugo.pace@dfinity.org>