fix: do not filter out test_suite targets in targets.py (#7208)
The recent master breakage on
https://github.com/dfinity/ic/commit/2047882f9eb3a37e3ad974b85e5a7298e62e6ddc
was because on the PR of that commit the target
`//ic-os/tests:integration_tests` was not tested. This is because we
filter out all targets except the ones of which their kind matches
`.*_tes`. Apparently test_suite targets don't match that kind so we need
to filter for them explicitly.