chore: simplify bazel uuid generation (#7169)
This avoids the need for an external python script to generate build
UUIDS and avoids the need to specify `--invocation_id`. Instead, this
uses a `BUILD_UUIDS` environment variable as a stack of UUIDs that the
`bazel` action can pop UUIDs from.

Our build containers do not include `uuidgen` so we do still use
`python` to generate the UUIDs.