refactor(P2P): Create dedicated Advert type (#7168)
This PR creates a dedicated advert type.

Currently, an Advert contains only the `StateSyncArtifactId`, which is
an id indicating, which state to sync.
In the future, we want to add additional information into the advert,
such as `partial_states`.

This PR wraps the `StateSyncArtifactId` into an outer `Advert` struct,
that we can add new fields to in the future.