chore(PocketIC): drop agent timeout when creating HTTP gateway (#7249)
This PR drops an extra timeout of 10s for initializing an agent when
creating HTTP gateway in PocketIC. This is supposed to mitigate test
[flakiness](https://github.com/dfinity/ic-testnet/actions/runs/18427697824/job/52515459984)
on busy CI runners.

Note that there's still the default timeout of 30 seconds imposed
implicitly by the `reqwest` crate making the HTTP connection. (This is
also why we need to bump that default timeout in a test so that the test
does not hit its own default timeout of 30s, but rather waits until the
default timeout of 30 seconds is reached inside PocketIC.)