perf(idkg): CON-1594 Use multi-threading in to create IDKG transcripts (#7158)
This PR moves the creation of the rayon thread pool out of the validator
and into the parent consensus module. This allows us to pass the same
thread pool also to the block maker.

The thread pool is subsequently used to create new IDKG transcripts in
parallel. To simplify the code, instead of adding parallelism to the
pre-signature state machine, we pre-compute all pre-signature
transcripts that are currently in creation in a separate step.
Subsequently, the computed transcripts are passed to the pre-signature
state machine, where they are consumed to advance the pre-signature
creation.

A previous approach instead envisioned moving the transcript creation
out of the payload builder and into the client thread (to reduce
pressure on the finalization rate). However, performance tests reveal
that doing so results in slightly worse signature throughput (though at
slightly higher finalization rate). The reason is likely because
creating the transcripts as part of the client thread is significantly
less efficient when considering the subnet as a whole, since each
creation is replicated by up to 34 nodes.

In the approach of this PR each transcript instead continues to be built
by the blockmaker exclusively. We can still control the pressure on the
finalization rate, by configuring the number of threads and the number
of pre-signatures to work on per block.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>