chore: add boot up time metrics (#7236)
# NODE-1602

Adding a simple bash script that runs different `systemd-analyze`
commands, parses their output and creates custom metrics to report
startup times of the nodes: it reports the time spent in the kernel and
userspace at boot up and it exports the time it took to initialize the
10 "slowest" units.

The script runs once, 10mins after the boot just to make sure it only
runs after the startup finished.

This should be useful to track the upgrade times.

It is a bit fragile as it relies on parsing the output of
systemd-analyze. If that changes, the metric breaks.