feat: [EXC-1848] Enrich `CallOrigin` with method name 2/2 (#7072)
This PR adds the name of the original method called to the `CallOrigin`
of a `CallContext`. This is useful when debugging but the field has no
other use.

Must not be merged before https://github.com/dfinity/ic/pull/7088 is
rolled out on all subnets.

### Technical Details

This PR changes some Protobuf definitions: `pb::CallOrigin` has an
additional variant `pb::CallOrigin::UserQuery(user_id, method_name)`
which will be used in place of the old `pb::CallOrigin::Query(user_id)`.
The other variants contain structs which have an additional field
`method_name`.

Since `method_name` is a Protobuf string which defaults to "" and the
field is purely decorative, i.e., not used in any APIs, there is no
additional `optional` indirection: We are fine with missing data being
represented as "".

The Rust struct `CallOrigin`'s variants are enhanced with `method_name`
which is not being used anywhere, but is displayed in error strings via
the `Debug` trait.

#### Backward compatibility:
New replica reads old pb data: 
Missing strings will default to "". These `CallOrigins` will be printed
without any useful information, but there is no regression. Newly
created `CallOrigins` are populated with `method_name` and that data is
written to new pb files. Furthermore, the old `CallOrigin` variant is
read to the new memory representation. When writing back to pb, the new
variant is written.

After a checkpoint, all pb data has the new format. 

#### Forward compatibility:
Old replica reads new pb data:
[This PR ](https://github.com/dfinity/ic/pull/7088)makes sure that old
replicas decode the new pb variant to the then-known Rust variant by
ignoring the new field.
This PR can be merged _after_ the above PR is rolled out on all subnets.

---------

Co-authored-by: Michael Weigelt <michael.weigelt@dfinity.com>