fix(nns-recovery): adjust NNS recovery testnet (#6873)
This PR adjusts the NNS recovery testnet with the following changes:
- Since
[#4108](https://github.com/dfinity/ic/pull/4108/files#diff-70eb215bc67459b41de730227aff0ea9d0593cc34feaea68480e5cdb35ef62a7R160),
the SHA256SUMS file does not have an asterisk in front of the digest. We
adapt `ic-recovery` with this new format.
- Similarly to the NNS recovery tests, we add our SSH key as `backup`
access to all nodes in the subnet, to train the recovery coordinator to
handle two keys. We thus move the relevant code in its own function
`grant_backup_access_to_all_nns_nodes` and call it at the two sites.
- In the system tests, we
[also](https://github.com/dfinity/ic/blob/a826c3d0f291cdceb7072c0ca61eb4f40ad4953f/rs/tests/nested/nns_recovery/common.rs#L306-L324)
disable admin access on all nodes except the DFINITY-owned one but I
don’t think we should do that here. Indeed that way, we can avoid having
to rebootstrap the testnet if we ever need to SSH into a node with
`admin` access for quick debugging or manual fixes. LMK if you think
otherwise.
- We suggest the user a working GuestOS version/url/hash to upgrade to.
- We display the mapping `VM hostname <-> HostOS/GuestOS IPs`.
- We add A Prometheus VM and Grafana dashboards to have some live
visualization
- The user can give some arguments to instruct the testnet to break a
given number of nodes at a given height to avoid having to manually
break a potentially high number of nodes.
- The default timeout is increased from 30 minutes to 90 minutes.