perf(crypto): CRP-2926 adapt ThresholdEcdsaSigInputs to only hold references (#7224)
Adapts the `ThresholdEcdsaSigInputs` to only hold references instead of
owning the data.

This is done because (1) it is conceptually cleaner and (2) it
eliminates the clonings required in the Consensus layer (in
`build_signature_inputs`, for nonce, caller, derivation path) to call
the tECDSA-related methods in the Crypto Component, i.e.,
create_sig_share, verify_sig_share, combine_sig_shares,
verify_combined_sig.

Note, however, that the savings are currently minor because the large
structs (`matched_data.pre_signature` and `matched_data.key_transcript`)
are contained in `Arc`s, but with this PR we could consider/evaluate
getting rid of the `Arc`s.

The `ThresholdEcdsaSigInputs` no longer use the (owned)
`ExtendedDerivationPath` nor the (owned) `Randomness` but rather
reference the raw underlying data as it is present in the
`SignWithThresholdContext`. This is done so that we don't need an
additional intermediate struct in the consensus layer that partially
holds owned data. Also, the `ThresholdEcdsaSigInputs` no longer
implements `Deserialize, Serialize` because `serde::Deserialize` cannot
be implemented on such reference structs, but luckily that wasn't needed
anyway.

In the tests, an owned variant (`ThresholdEcdsaSigInputsOwned`) was
created for improved usability in tests so that testing functions can
create and return threshold sig inputs (e.g.,
`generate_tecdsa_protocol_inputs` and `build_signature_inputs`).