fix: don't publish orchestrator on macOS
This filters out the orchestrator binary from macOS Darwin artifacts.

The orchestrator is never used on macOS and it involves building many
mostly-Linux-specific crates — we might as well not build it.

Ideally the implementation would be more explicit by specifying Linux
and macOS targets separately, but we can iterate on this easily.