chore(Ledger_suite): FI-1888: Remove ledger_suite dependency on ic-nns-constants (#7267)
Remove the `rs/ledger_suite/icp` dependency on `ic-nns-constants`. The
aforementioned crate is only used to get the NNS canister IDs, but it
now transitively pulls in e.g., `prometheus`. These constants will not
change in the foreseeable future, so this PR proposes copying them to
the `icp-ledger` crate (which is already a dependency in all the crates
that previously depend on `ic-nns-constants`) and removing the
`ic-nns-constants` dependency.