fix: Turn off query stats for transform functions of https outcalls (#7144)
Tracking query stats for transform functions related to https outcalls
was inadvertently added in a previous refactoring.

This does not make much sense as these calls can be fully replicated or
non-replicated (especially with the flexible outcalls feature that is
coming up soon). Additionally, these calls will be charged for
separately with the revamp of the https outcalls feature and thus,
tracking them in query stats which would be used to (eventually) charge
cycles for regular queries is also not something that would make sense.

The proposed solution is to disable query stats tracking for the https
outcalls service that handles the execution for transform query
functions.