chore(nns-recovery): higher timeout when waiting for subnet membership change (#7281)
Spawning up the testnet with the current `master` for today's drill
actually failed because it was the first time it was launched with the
production DKG interval of 499. The driver would
[timeout](https://dash.zh1-idx1.dfinity.network/invocation/b6072033-5fb2-48db-8a46-55a0e7a954de#@2343)
when waiting for the subnet membership to change from one node to the N
virtualized nodes. Indeed, the subnet membership is updated at summary
heights, so a longer DKG interval would mean that the membership change
would take more time to be observed.

Bumping from 500 seconds ( == `READY_WAIT_TIMEOUT`) to 1000 seconds
[worked](https://dash.zh1-idx1.dfinity.network/invocation/f7414a9d-71b1-41b0-9d98-8d7f7ceb1b65#@1587)
(see how it took almost 600 seconds). This PR bumps it to 900 (15
minutes). This number is quite overkill for the system tests where the
interval is 9, but I think having to wait 400 additional seconds before
observing a flake isn't that big of a deal, while it is great to be able
to cover higher DKG interval lengths (especially the production one).

It also uses a consistent style across the file (`Duration::from_secs`
-> `secs`)