chore: Fix flaky `test_stable_registry_version` (#7290)
This test uses `std::thread::sleep` and implicitly makes assumptions
about the amount of time passed after each sleep. This leads to
occasional flakiness if the test sleeps longer than expected.

It would require quite some refactoring to mock the `current_time()`
calls in this test (as they are done by the blockmaker and the registry
client). However, we can avoid the flakiness by not making assumptions
about the length of each sleep. This is achieved by storing the
timestamp of each registry version update in a variable, and using it to
calculate the precise registry version age.