refactor: Use compute capacity based on scheduler cores in tests (#7188)
This refactors how compute capacity is calculated in `ExecutionTest`.
Instead of hardcoding a value of 100 (which maps to `scheduler cores ==
2`), it uses the `RoundSchedule::compute_capacity_percent` function
similar to production code.

Additionally, it introduces an explicit check that when `scheduler_cores
== 1` in tests (which is a useful scenario to test various cases of
subnet's available memory as the math is simpler), DTS is disabled by
setting the slice limits to much message limits.

Furthermore, it ties the `resource_saturation_scaling` option in tests
with the number of scheduler cores again matching what's going on in
production.

Finally, the subnet available memory is also calculated the same way as
in production instead of using `SubnetAvailableMemory::new_for_testing`
which does not perform the correct scaling based on the number of
scheduler cores.

All these changes pave the way for a follow up PR which will refactor
how `ExecutionTest` sets up the execution services and will match what
happens in production to further consolidate things and ensure testing
is as accurate as possible wrt production setups.

---------

Co-authored-by: Martin Raszyk <martin.raszyk@dfinity.org>