refactor: Use u64 for methods that set capacities in tests (#7306)
This PR refactors a few functions in the `ExecutionTestBuilder` to
accept `u64` instead of `i64`. This makes more sense as these amounts
represent bytes and it results in fewer casts being needed in the tests.
There are still some that are needed in cases where we need to calculate
some available memory which is an `i64` but overall this is still a
change that cleans up some of the code.