refactor: Use setup_execution consistently across production and test code (#7138)
This PR refactors some test code paths to re-use the `setup_execution`
function to construct the various execution services needed. This
ensures that we don't have code duplication (at the moment the same
(roughly) code was repeated 3 times in `ExecutionTest`, `SchedulerTest`
and execution benches).

This is achieved by defining an `ExecutionServicesForTesting` struct
which captures what's needed for test purposes. There are some
differences compared to the production version which made it easier to
have a separate struct. Specifically, the query handler is preferred to
be exposed as the sync internal query handler to allow the tests to
remain sync as well whereas the production environment needs an async
one. Another difference is that the test services skip returning the
scheduler as we only need it in `SchedulerTest` and they require to have
access to the (crate only visible) `SchedulerImpl` as opposed to the
only public component which is the `Scheduler` trait (making
`SchedulerImpl` public for this seemed unnecessary). `SchedulerTest`
instead creates its own scheduler as it used to but based on the
services returned by `ExecutionServicesForTesting::setup_execution`.

As a nice side-effect of this refactoring, we can also make `pub(crate)`
some other parts like the Hypervisor's methods or the
`InternalQueryHandler` constructor as they are now only constructed by
`setup_execution` and not directly by any other test code outside the
`execution_environment` crate.