feat(ICRC-Index): FI-1759: stop indexing and report error in case of unknown block (#6996)
This PR enhances the index block validation by additionally checking for
additional unrecognised fields. If an unknown/incorrect blocks is
detected, we don't trap anymore but stop the indexing timer and log the
error. The error can be accessed via the `/logs` http endpoint, e.g. for
the ckBTC index: https://n5wcd-faaaa-aaaar-qaaea-cai.raw.icp0.io/logs.

Since the cycles ledger and the latest iteration of the ICRC-3 standard
allow for fees in `Mint` and `Burn` operations, we additionally add
these fields in the respective ICRC ledger operations - without those
fields the index would not be able to sync the cycles ledger blocks.
These fields are however never used or populated by the ICRC ledger -
`Mint` and `Burn` transactions with positive fees are always rejected.