feat(canister_logging): charge for fetching logs via update call (#7033)
This PR implements charging for fetching canister logs for update calls
(when feature is enabled).

The actual fees are just placeholders for now, should be adjusted after
benchmarking the actual costs.

When charging for logs user is required to provide enough cycles to
cover a full response 2MB, but it only charges the actual amount of
bytes returned.