fix: experiment to stabilize //rs/bitcoin/adapter:adapter_integration_test (#7326)
The `//rs/bitcoin/adapter:adapter_integration_test` is very flaky. The
hypothesis is that this is caused by failing HTTP connections to the
bitcoind server which might be caused by an overloaded runner. So to
reduce load we reduce the number of parallel tests to 4 by setting
`RUST_TEST_THREADS` to 4 and by giving the test 4 CPUs instead of just
1.