test(nns-recovery): [CON-1601] fix the DFINITY-owned node like the other NPs (#7184)
Up until now, it was actually impossible not to fix the DFINITY-owned
node with `ic-recovery` using it the "normal way", i.e. without
specifying CLI arguments. Indeed, the `UploadCUPAndRegistry` was
mandatory because it used the same node as the one we upload the state
to (which is always set). (We could still have skipped the step with
`--skip UploadCUPAndRegistry`)

This is a regression from this
[commit](https://github.com/dfinity/ic/commit/08012e9a783cec2af4dca8bdab57a8fba48d233c),
where we could previously skip the `UploadCUPAndRegistry` by not
inputting anything and later input the `upload_method` for
`UploadState`. Since the commit swapped both steps, it is not possible
to ignore `UploadCUPAndRegistry` anymore.

This PR uses `wait_for_cup_node` for `UploadCUPAndRegistry`, which also
makes more sense as this should be the same variable as the `WaitForCUP`
step. To prevent this regression from happening again, this PR also
modifies the existing `:nr_no_bless` system test to also fix the
DFINITY-owned node like the other NPs, i.e. not through `ic-recovery`
but with the `guestos-recovery-upgrader`.