feat: Consider IDKG transcripts stored in state during complaints phase (#7214)
IDKG complaints and openings are created/verified based on "active
transcripts" (i.e. transcripts that are part of available pre-signatures
or key transcripts). If a received complaint/opening doesn't correspond
to such an active transcript, it is discarded instead. Previously, these
active transcripts were always referenced by the finalized tip.
Therefore, returning the active transcripts referenced by the finalized
tip was enough to determine if an incoming complaint/opening should be
handled.

In the future, we will start storing pre-signature transcripts and some
key transcripts in the replicated state exclusively. Regardless, we
should still be able to handle complaints/openings for these
transcripts.

This PR therefore changes the "active transcripts" used by the
complaints handler, to also consider:
1. Avaliable pre-signatures that were delivered in blocks below the
finalized tip, that haven't been executed yet
2. Pre-signature transcripts that are stored in the state's
pre-signature stash
3. Key transcripts that are stored in the state's pre-signature stash
4. Key and pre-signature transcripts that were paired to ongoing
signature requests.

This requires us to hand the state reader to the complaints handler
during initialization.

Note that before this PR, complaints/openings corresponding to a
transcript ref that cannot be resolved were removed from the unvalidated
pool with `HandleInvalid`. After this PR, transcripts that cannot be
resolved are no longer part of the "active transcripts". Therefore,
these complaints/openings are instead removed using `RemoveUnvalidated`.
Personally, I think this actually makes more sense, because whether or
not a transcript ref can be resolved doesn't depend on the received
complaint/opening, and rather indicates a bug in the system. The only
difference between the two change actions is that `HandleInvalid` will
raise an alert. However, a transcript failing to be resolved similarly
raises a metric, which I plan to connect to an alert in a future PR.

Additionally, the following problem might occur: The complaints handler
only has access to the finalized chain, which consists of all finalized
blocks at and above the latest summary. However, during every interval,
there might be a short period of time, where the latest certified state
height is below the finalized summary height. In that case, we would
wrongfully exclude all pre-signature transcripts that were delivered in
blocks between the latest certified height and the summary.
Consequently, complaints/openings received for these transcripts during
that time would be wrongfully purged forever. To avoid the problem, this
PR ensures to only call the complaints handler if the certified state
height is part of the current finalized chain.