feat(registry): Invariant check to prevent unassigned nodes from having ssh_write_access (#7314)
This prevents a node from having `ssh_node_state_write_access` when it
is not currently part of a subnet. This field should only be set during
the subnet recovery process, at which point the node with this field set
should not be removed from the subnet. If it were removed, it could in
theory be added to a different subnet, giving the party possessing the
SSH key access to a different subnet's data. In theory, if enough nodes
were able to have this field set in a subnet, the subnet could be
controlled by the party with the ssh keys.