refactor(canlog): Use `regex-lite` instead of `regex` for parsing regular expressions (#7353)
Change the regex engine in the `canlog` library to use
[`regex-lite`](https://crates.io/crates/regex-lite) instead of
[`regex`](https://crates.io/crates/regex) to improve the binary size.

By default the `regex` crate has a lot of enabled features that lead to
a fairly large binary once compiled. From the
[docs](https://github.com/rust-lang/regex/blob/5ea3eb1e95f0338e283f5f0b4681f0891a1cd836/src/lib.rs#L1140C1-L1144C22)
> This means that there is a lot of
code dedicated to performance, the handling of Unicode data and the
Unicode
data itself. Overall, this leads to more dependencies, larger binaries
and
longer compile times

BREAKING CHANGE: The error type in the public methods
`RegexString::compile` and `RegexString::try_is_valid` is changed to no
longer leak the error type of the underlying library.