chore: One more (f)stab at UpgradeVM systemd config (#7355)
NODE-1745

A more robust and faster approach to skip trying to mount the encrypted
partitions in UpgradeVMs where these partitions are not available. The
previous approach of just waiting until the timeout is over was slow and
the 30s timeout was too low causing flakiness. Furthermore, `nofail` had
to be added to the mounts.

This new approach adds a new `maybe-disable-store-mounts.service` which
disables the encrypted mounts when the Guest is running as an Upgrade
VM.

This PR also adds a new path under `/run/config/guest_vm_type/` where a
file named `upgrade` or `default` is created showing the VM type. This
can be used in path conditions (so not only in service units like exec
conditions).