fix: dep-scan trivy file not found error (#7368)
When scanning with trivy, it fetches the vuln DB directly from github
which is very flaky. Previously it was guaranteed that the scan results
file was always written and had a zero size when fetching the vuln DB
failed. However this seems to have changed recently. Now it may also be
the case that the file is not created at all. In this PR we handle this
case. Also we increase the maximum number of scan attempts to 50.