feat(ckdoge): endpoints to retrieve events, metrics, dashboard and logs (#7280)
Add 2 standard endpoints to retrieve information related to the canister
operation:

1. `get_events`: retrieve log of state changes, similar to that of the
ckBTC minter. Contrary to the ckBTC minter, the endpoint is currently
hidden since it reuses for now the same events as the ckBTC minter,
although some are not applicable to the ckDOGE minter.
2. `http_request`: hidden endpoint to contact the canister via HTTP
request to retrieve logs, metrics and the dashboard. To be able to reuse
the same implementation as for the ckBTC minter the code was moved from
`ckbtc/minter/src/main.rs` to `ckbtc/minter/src/queries.rs`.