feat: Implement script for quick reloading of the HostOS
The reload_hostos.sh allows reloading the HostOS with a new version. The script is packaged with its dependencies into a self-extracting tar. This tar will be SCP-d over to a running bare-metal instance from tests. This PR also contains a few modifications to existing scripts to allow the new usecase.