feat(Ledgers): FI-1881: Check ledger liquid cycles balance before spawning archive (#7363)
Check the ledger liquid cycles balance before attempting to spawn an
archive.

If the ledger calls `create_canister` and tries to attach more than its
current liquid cycles balance, the call will trap, rolling back the
transaction that triggered the archive creation. This leads to a
situation where a ledger cannot process more transactions until it has
been topped up.

Some checks are added to make sure that the configured
`cycles_for_archive_creation` is enough to cover the cost of canister
creation, plus the installation and some initial operations of the
archive canister. A check is also added to make sure that the ledger
still has some minimum amount of liquid cycles left after spawning the
archive. In case any of the checks fail, an error is logged, and a
metric counter is incremented. This allows ledger operators to add
Prometheus alerts based on the metric, and to determine the cause of the
error using the logs.