refactor(networking): Manage chunks to download explicitly (#7343)
This PR changes the way chunks to download are managed in the ongoing
state sync handler.

This PR is mainly a refactor, but there is a slight change in the logic,
so it is a functional change.

## Backgound (current situation)

The `OngoingStateSync` is a manager that gets spawned when a state sync
is started. It's job is to process incoming adverts, and spawn downloads
on peers, then deliver the downloaded chunks to state sync.

As part of this job, it occasionally calls `get_chunks_to_download` and
collects the chunk id's into a `Box<dyn Iterator<Item = ChunkId> +
Send>`, on which it calls `.next()` to get a new chunk id to spawn a
download.
If there are no chunks in the iterator, it calls
`get_chunks_to_download` again, which may fill up the iterator.
This loop is only limited by the number of available downloads, so we
might loop multiple times through it without doing anything.


On success, it delivers the downloaded chunks to state sync, on
`NoContent`, it removes the state. On other errors, it simply drops the
download.

This means, that if a download fails (which actually happens quite
regularly due to network exhaustion), the chunk id is basically
"forgotten", until we run out of chunks to download, and call
`get_chunks_to_download` again.

## Changes

The current behaviour is slightly inefficient, because we loop
unnecessarily in `spawn_chunks_download` and we call
`get_chunks_to_download` more often than we need.

This PR introduces an explicit struct `ChunksToDownload`, which manages
the chunks to download list on behalf on `OngoingStateSync`. Rather than
forgetting a chunk id on an error, we call
`ongoing_downloads.download_failed` to re-queue the chunk for download.

Furthermore, on `spawn_chunks_download`, we only check once, if we need
to call `get_chunks_to_download` and fill up our list of chunks. This
now only happens when we are finished with downloading one layer (i.e.
when we finished downloadin manifest chunks and start downloading
chunks), since failed downloads are re-queued by the ongoing state
manager directly.

The behaviour is currently based on a `VecDequeue`, but can be changed
when introducing Fast State Sync with IC-1502.

---------

Co-authored-by: Leo Eichhorn <99166915+eichhorl@users.noreply.github.com>
Co-authored-by: kpop-dfinity <125868903+kpop-dfinity@users.noreply.github.com>